requirements.txt
numpy
rosbags
bagpy
opencv-python
opencv-contrib-python
scikit-learn
cvbridge3
pyyaml
tk
matplotlib
--extra-index-url https://rospypi.github.io/simple/ 
roslz4